Created a GameModel class.
Added Constructor that creates a GameGrid with X-amount of columns and Y-amount of rows.
Added two functions to get the state of a tile in the gameGrid, one that takes a row and a column as input, and one that just takes a position and then calculates the row and column.
Added two functions to set the state of a tile in the gameGrid, its built on the same principle as above.
Added two functions to get the row and column from a position